Some refactoring and added a reference to the parent_key for regexes
