+++ napalm_base/base.py
@@ -26,13 +26,10 @@ import napalm_base.helpers
 import napalm_base.constants as c
 
 from napalm_base import validate
+from napalm_base import yang_helpers
 
 
 import napalm_yang
-
-import yaml
-
-import re
 
 
 class NetworkDriver(object):
@@ -1512,101 +1509,8 @@ class NetworkDriver(object):
 
     def oc_populate_interfaces(self):
         self.interfaces = napalm_yang.oc_if.Interfaces().interfaces
-
-        with open("napalm_eos/openconfig_mappings.yaml", "r") as f:
-            oc_mappings = yaml.load(f.read())
-
         config = self._oc_all_config()
-
-        self._oc_parse_text(self.interfaces, config, config, oc_mappings["interfaces"])
-
-    def _oc_parse_text(self, model, config, partial_config, mappings):
-        for k, v in model.items():
-            if issubclass(v.__class__, napalm_yang.List):
-                self._oc_parse_text_to_list(v, config, mappings[k])
-            elif k == "state":
-                continue
-            elif issubclass(v.__class__, napalm_yang.BaseBinding):
-                self._oc_parse_text(v, config, partial_config, mappings[k])
-            else:
-                self._oc_parse_text_to_attr(v, partial_config, mappings[k])
-
-    def _oc_parse_text_to_list(self, model, config, mappings):
-        block_matches = re.finditer(mappings["_block_capture"], config, re.MULTILINE)
-
-        for match in block_matches:
-            name = match.group("key")
-            block_config = match.group("block")
-
-            if "_parent_key" in mappings.keys():
-                parent_key = re.search(mappings["_parent_key"], block_config,
-                                       re.MULTILINE).group("parent")
-                path = mappings["_path"].format(parent_key, name)
-                obj = eval("self.{}".format(path))
-            else:
-                obj = model.get_element(name)
-
-            self._oc_parse_text(obj, config, block_config, mappings)
-
-    def _oc_parse_text_to_attr(self, attr, config, mappings):
-        match = re.search(mappings["_search"], config, re.MULTILINE)
-
-        if mappings["_type"] == "boolean":
-            attr(match is not None)
-            return
-
-        if match:
-            if mappings["_type"] == "mapping":
-                value = mappings["_map"][match.group("value")]
-            else:
-                value = match.group("value")
-        else:
-            value = mappings["_default"]
-
-        try:
-            attr(value)
-        except ValueError:
-            attr(eval(value))
+        return yang_helpers.TextExtractor().populate(self.interfaces, config)
 
     def translate_yang_model(self, model):
-        with open("napalm_eos/openconfig_translation.yaml", "r") as f:
-            translation_map = yaml.load(f.read())["interfaces"]
-
-        return self._translate_yang_model(model, translation_map)
-
-    def _translate_yang_model(self, model, translation_map):
-        translation = ""
-        for k, v in model.items():
-            if issubclass(v.__class__, napalm_yang.List):
-                translation += self._yang_translate_list(v, translation_map[k])
-            else:
-                if issubclass(v.__class__, napalm_yang.BaseBinding):
-                    if not v._meta["config"]:
-                        continue
-                    translation += self._translate_yang_model(getattr(model, k), translation_map[k])
-                translation += self._yang_translate_string(translation_map[k].get("_string", ""),
-                                                           value=getattr(model, k))
-                translation += self._yang_translate_map(translation_map[k].get("_map", {}),
-                                                        value=getattr(model, k))
-        return translation
-
-    def _yang_translate_list(self, l, translation_map):
-        translation = ""
-        for element_key, element_data in l.items():
-            translation += self._yang_translate_string(translation_map.get("_string", ""),
-                                                       key=element_key)
-            translation += self._translate_yang_model(element_data, translation_map)
-
-        return translation
-
-    def _yang_translate_map(self, m, value, **kwargs):
-        if m:
-            return self._yang_translate_string(m[str(value)]["_string"], value=value, **kwargs)
-        else:
-            return ""
-
-    def _yang_translate_string(self, string, **kwargs):
-        if string:
-            return string.format(**kwargs)
-        else:
-            return ""
+        return yang_helpers.TextTranslator().translate(model)
